pyproject.toml
[project]
name = "trm-sudoku-agi-model"
version = "0.1.0"
description = "Tiny Reasoning Model for Sudoku-Extreme puzzles"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.2.0",
    "datasets>=2.18.0",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "pyyaml>=6.0.1",
    "tqdm>=4.66.0",
    "requests>=2.31.0",
]

[project.scripts]
trm-train = "scripts.train:main"
trm-evaluate = "scripts.evaluate:main"
trm-compare-llm = "scripts.compare_llm:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
